Add container for the smb net share
This is simple container with guest access configured. See smb.conf for
the details. smbuser:smbuser is created and can be used for the testing
with authentication, for that disable guest access in the config file.

Mounting can be done with the following command:
```
mount_smbfs -N //guest@192.168.99.101/Shared ./smb
```